+++ react/components/ConditionsSelector/ScopeSelector.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
-import { injectIntl } from 'react-intl'
-
-import Dropdown from '../form/Dropdown'
+import { InjectedIntl, injectIntl } from 'react-intl'
+import { Dropdown } from 'vtex.styleguide'
 
 const SITE_SCOPE_CONDITION = {
   label: 'pages.conditions.scope.site',
@@ -18,6 +17,14 @@ const SCOPE_CONDITIONS = [
     value: 'route',
   },
 ]
+
+const getOptions = ({intl, shouldEnableSite}: {intl: InjectedIntl, shouldEnableSite: boolean}) => {
+  const conditions = shouldEnableSite ? [...SCOPE_CONDITIONS, SITE_SCOPE_CONDITION] : SCOPE_CONDITIONS
+  return conditions.map(option => ({
+    ...option,
+    label: option.label && intl.formatMessage({id: option.label}),
+  }))
+}
 
 interface Props {
   onChange: (value: ConfigurationScope) => void
@@ -36,11 +43,7 @@ const ScopeSelector = ({
       id: 'pages.editor.components.conditions.native.label',
     })}
     onChange={onChange}
-    options={{
-      enumOptions: shouldEnableSite
-        ? [...SCOPE_CONDITIONS, SITE_SCOPE_CONDITION]
-        : SCOPE_CONDITIONS,
-    }}
+    options={getOptions({intl, shouldEnableSite})}
     value={value}
   />
 )