+++ react/components/form/ArrayFieldTemplateItem.tsx
@@ -1,12 +1,17 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
-import { SortableElement, SortableHandle } from 'react-sortable-hoc'
-import { Transition } from 'react-spring'
-
+import { ArrayFieldTemplateProps } from 'react-jsonschema-form'
+import {
+  SortableElement,
+  SortableElementProps,
+  SortableHandle,
+} from 'react-sortable-hoc'
+import { animated, Transition } from 'react-spring'
 import DragHandle from '../icons/DragHandle'
 import TrashSimple from '../icons/TrashSimple'
+import SimpleFormattedMessage from './SimpleFormattedMessage'
 
-const stopPropagation = fn => e => {
+const stopPropagation = (fn: (e: React.MouseEvent) => void) => (e: React.MouseEvent) => {
   e.stopPropagation()
   return fn(e)
 }
@@ -15,59 +20,40 @@ const Handle = SortableHandle(() => (
   <DragHandle size={12} className="accordion-handle" />
 ))
 
-class ArrayFieldTemplateItem extends Component {
-  static propTypes = {
+interface IProps {
+  children?: React.ReactElement<{ formData: number }>
+  formIndex: number
+  hasRemove: boolean
+  isOpen: boolean
+  onClose: () => void
+  onOpen: (e: React.MouseEvent) => void
+  showDragHandle: boolean
+  schema: any
+}
+
+interface State {
+  autoHeight: boolean
+}
+type PropsFromItemTemplateProps = Pick<
+  ArrayFieldTemplateProps['items'][0],
+  'onDropIndexClick' | 'hasRemove' | 'children'
+>
+type Props = IProps & SortableElementProps & PropsFromItemTemplateProps
+
+class ArrayFieldTemplateItem extends Component<Props, State> {
+  public static propTypes = {
     children: PropTypes.node,
     formIndex: PropTypes.number,
     hasRemove: PropTypes.bool,
-    onDropIndexClick: PropTypes.func,
-    schema: PropTypes.object,
     isOpen: PropTypes.bool,
-    onOpen: PropTypes.func,
     onClose: PropTypes.func,
+    onDropIndexClick: PropTypes.func,
+    onOpen: PropTypes.func,
+    schema: PropTypes.object,
     showDragHandle: PropTypes.bool,
   }
 
-  state = {
-    autoHeight: false,
-  }
-
-  handleLabelClick = e => {
-    const { isOpen, onOpen, onClose } = this.props
-
-    if (isOpen) {
-      onClose(e)
-    } else {
-      onOpen(e)
-    }
-  }
-
-  handleOnRest = () => {
-    if (!this.state.autoHeight) {
-      this.setState({ autoHeight: true })
-    }
-  }
-
-  handleOnStart = () => {
-    if (this.state.autoHeight) {
-      this.setState({ autoHeight: false })
-    }
-  }
-
-  renderChildren = styles => {
-    const { children } = this.props
-    return (
-      <div
-        style={{
-          ...styles,
-          height: this.state.autoHeight ? 'auto' : styles.height,
-        }}>
-        {children}
-      </div>
-    )
-  }
-
-  render() {
+  public render() {
     const {
       children,
       schema,
@@ -86,17 +72,21 @@ class ArrayFieldTemplateItem extends Component {
       <div
         className={`accordion-item bb b--light-silver ${
           showDragHandle ? '' : 'accordion-item--handle-hidden'
-        }`}>
+        }`}
+      >
         <div className="accordion-label" onClick={this.handleLabelClick}>
           <div className="flex items-center">
             {showDragHandle && <Handle />}
-            <label className="f6 accordion-label-title">{title}</label>
+            <label className="f6 accordion-label-title">
+              <SimpleFormattedMessage id={title} />
+            </label>
           </div>
           <div className="flex items-center accordion-label-buttons">
             {hasRemove && (
               <button
                 className="accordion-icon-button accordion-icon-button--remove"
-                onClick={stopPropagation(onDropIndexClick(formIndex))}>
+                onClick={stopPropagation(onDropIndexClick(formIndex))}
+              >
                 <TrashSimple size={15} />
               </button>
             )}
@@ -105,20 +95,36 @@ class ArrayFieldTemplateItem extends Component {
         <div
           className={`accordion-content ${
             isOpen ? 'accordion-content--open' : ''
-          }`}>
+          }`}
+        >
           <Transition
-            keys={isOpen ? ['children'] : []}
+            native
+            config={{ duration: 300 }}
+            items={isOpen ? ['children'] : []}
             from={{ opacity: 0, height: 0 }}
             enter={{ opacity: 1, height: 'auto' }}
             leave={{ opacity: 0, height: 0 }}
-            onRest={this.handleOnRest}
-            onStart={this.handleOnStart}>
-            {isOpen ? [this.renderChildren] : []}
+          >
+            {this.renderChildren}
           </Transition>
         </div>
       </div>
     )
   }
+
+  private handleLabelClick = (e: React.MouseEvent) => {
+    const { isOpen, onOpen, onClose } = this.props
+
+    if (isOpen) {
+      onClose()
+    } else {
+      onOpen(e)
+    }
+  }
+
+  private renderChildren = (item: string) => (styles: React.CSSProperties) => (
+    <animated.div style={styles}>{this.props.children}</animated.div>
+  )
 }
 
 export default SortableElement(ArrayFieldTemplateItem)