+++ react/components/form/MultiSelect.tsx
@@ -1,14 +1,25 @@
-import React, { Fragment } from 'react'
-import ReactSelect from 'react-select'
 import PropTypes from 'prop-types'
+import React, { Fragment } from 'react'
+import ReactSelect, { Option } from 'react-select'
 
-const getChangeHandler = onChange => value => {
-  const formattedValue = value.map(item => item.value)
-
-  return onChange(formattedValue)
+interface Props {
+  autofocus?: boolean
+  disabled?: boolean
+  id?: string
+  label?: string
+  onChange: (newValue: string[]) => void
+  options: {
+    enumOptions: Option[]
+  }
+  placeholder: string
+  schema?: {
+    disabled?: boolean
+    title: string
+  }
+  value: string[]
 }
 
-const MultiSelect = ({
+const MultiSelect: React.SFC<Props> = ({
   autofocus,
   disabled,
   id,
@@ -23,16 +34,23 @@ const MultiSelect = ({
     {(label || (schema && schema.title)) && (
       <label>
         <span className="dib mb3 w-100">
-          <span>{label || schema.title}</span>
+          <span>{label || (schema && schema.title)}</span>
         </span>
       </label>
     )}
     <ReactSelect
       autoFocus={autofocus}
-      disabled={disabled || (schema && schema.disabled) || enumOptions.length === 0}
+      disabled={
+        disabled || (schema && schema.disabled) || enumOptions.length === 0
+      }
       id={id}
       multi
-      onChange={getChangeHandler(onChange)}
+      onChange={optionValues => {
+        const formattedValue = (optionValues as Option[]).map(
+          (item: Option) => item.value as string,
+        )
+        onChange(formattedValue)
+      }}
       options={enumOptions}
       placeholder={placeholder}
       value={value}
@@ -64,7 +82,7 @@ MultiSelect.propTypes = {
   placeholder: PropTypes.string,
   schema: PropTypes.shape({
     disabled: PropTypes.bool,
-    title: PropTypes.string.isRequired
+    title: PropTypes.string.isRequired,
   }),
   value: PropTypes.arrayOf(PropTypes.string),
 }