+++ react/components/form/SimpleFormattedMessage.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { FormattedMessage } from 'react-intl'
+
+const SimpleFormattedMessage: React.SFC<{
+  id: string;
+}> = ({ id }) => (<FormattedMessage id={id}>{txt => <>{txt}</>}</FormattedMessage>)
+
+export default SimpleFormattedMessage