+++ react/typings/render.d.ts
@@ -1,23 +1,25 @@
 /* Typings for `render-runtime` */
 declare module 'render' {
-  import {ReactElement, Component} from "react"
+  import { ReactElement, Component } from 'react'
 
   export const ExtensionPoint: ReactElement
   export const Link: ReactElement
   export const NoSSR: ReactElement
   export const RenderContextConsumer: ReactElement
   export const canUseDOM: boolean
-  export const withRuntimeContext: any
+  export const withRuntimeContext: <TOriginalProps extends {}>(
+    Component: ComponentType<TOriginalProps & RenderContextProps>,
+  ) => ComponentType<TOriginalProps>
 
-  interface RenderComponent<P={}, S={}> extends Component<P,S> {
+  interface RenderComponent<P = {}, S = {}> extends Component<P, S> {
     getCustomMessages?: (locale: string) => any
     schema: ComponentSchema
-    getSchema?: (any, any) => ComponentSchema
+    getSchema?: (any, any?) => ComponentSchema
     uiSchema: UISchema
   }
 
   export interface ComponentsRegistry {
-    [component: string]: RenderComponent<any,any>
+    [component: string]: RenderComponent<any, any>
   }
 
   export interface Window extends Window {