+++ react/typings/vtex.styleguide.d.ts
@@ -1,14 +1,22 @@
 declare module 'vtex.styleguide' {
-  import {ReactElement} from "react"
+  import { ReactElement } from 'react'
 
+  export const Alert: ReactElement
   export const Badge: ReactElement
   export const Button: ReactElement
   export const Card: ReactElement
+  export const Dropdown: ReactElement
   export const EmptyState: ReactElement
   export const IconArrowBack: ReactElement
+  export const IconEdit: ReactElement
   export const Input: ReactElement
+  export const Modal: ReactElement
+  export const PageHeader: ReactElement
   export const Pagination: ReactElement
+  export const Radio: ReactElement
   export const Spinner: ReactElement
+  export const Tab: ReactElement
   export const Table: ReactElement
+  export const Tabs: ReactElement
   export const Toggle: ReactElement
 }